Renamed and split MapRenderHelper for more convenient subclassing, set coroutine scope directly for offscreen rendering
