+++ android/src/main/java/io/openmobilemaps/mapscore/map/view/MapView.kt
@@ -39,7 +39,8 @@ import javax.microedition.khronos.egl.EGLConfig
 import javax.microedition.khronos.opengles.GL10
 
 open class MapView @JvmOverloads constructor(context: Context, attrs: AttributeSet? = null, defStyleAttr: Int = 0) :
-	GlTextureView(context, attrs, defStyleAttr), GLSurfaceView.Renderer, AndroidSchedulerCallback, LifecycleObserver, MapViewInterface {
+	GlTextureView(context, attrs, defStyleAttr), GLSurfaceView.Renderer, AndroidSchedulerCallback, LifecycleObserver,
+	MapViewInterface {
 
 	var mapInterface: MapInterface? = null
 		private set
